Bug 1189915 - Add privacy policy link to Settings
